Properly handle redirects in internal '_curl' function
Even though we told `curl` to follow 'Location' headers (see `curl -L`),
we didn't parse its output correctly. The status code we evaluated was
that of the first request, not the final one, and since we treat a 3xx
status code as error, this always failed. We still treat a 3xx status
code as failure, we even treat a 1xx status code as failure now, but we
now properly evaluate the status code of the final request.

We furthermore now only return the headers of the final request,
simplifying usage of the internal '_curl' function.

Note: `tail` needs to skip +2 lines to get `$BODY`, because Bash
collapses the trailing `\n\n` of `$HEADERS` to a single `\n`.